Extract GAN Backend
