+++ RenewableGANBackend/gan_utils.py
